Switch shp2pgsql over to use liblwgeom instead of its own internal parser. Some notes from looking at the code:
i) I've abstracted the I/O formatting functions into a new OutputGeometry() function since it allow the logic concerning output formatting to kept in one place, rather than sprinkled throughout all the other Insert* functions.

ii) InsertPoint() and InsertMultiPoint() have been combined, since it seems that the same code will work for both - all that is required is to know whether to produce a MULTIPOINT collecton or a POINT at the end.

iii) I've added additional comments within the Insert* functions to help clarify what is going on in places

iv) It appears some corner cases were missing within the conversion code with respect to Z/M coordinates, so in theory the new code should do a better job.

All in all, the source code is slightly reduced in size (although of course the binary is larger after linking with liblwgeom), and seems a lot more readable to my eyes. A quick TODO is to go through the file and replace the #if...#endif sections related to debugging with LWDEBUG(F) options.  

NOTE: I've had to change the Makefile to allow linking against the maths library and liblwgeom.a, so some manual intervention on the automated build may be required ;)



git-svn-id: http://svn.osgeo.org/postgis/trunk@3217 b70326c6-7e19-0410-871a-916f4a2858ee